+++ packages/app/src/analysis/stores/analysis-store.ts
@@ -1,4 +1,8 @@
-import { AnalysisType, OneOfValues, ProgrammingLanguage } from "shared-types";
+import {
+  AnalysisType,
+  OneOfValues,
+  ProgrammingLanguage,
+} from "@ct/shared-types";
 import { useStore } from "zustand";
 import { createStore } from "zustand/vanilla";