+++ packages/app/src/report/components/sbom/SBOMTable.tsx
@@ -1,8 +1,9 @@
+import { SbomPackage } from "@ct/shared-types";
 import { FC } from "react";
-import { SbomPackage } from "shared-types";
 import { ResponsiveTable, TableOptions } from "../../../common/ResponsiveTable";
 import { SeverityBadge } from "../../../common/SeverityBadge";
 import { useReportStore } from "../../stores/fe-report-store";
+import { SBOMTableNameCell } from "./SBOMTableNameCell";
 
 export const SBOMTable: FC = () => {
   const { packages } = useReportStore();
@@ -12,14 +13,20 @@ export const SBOMTable: FC = () => {
 const tableOptions: TableOptions<SbomPackage> = {
   emptyMessage: "No packages information found",
   cells: [
-    { label: "Name", key: "packageName" },
+    {
+      label: "Name",
+      cellRenderer: (row) => <SBOMTableNameCell pkg={row} />,
+    },
     { label: "Version", key: "packageVersion" },
     {
       label: "Severity",
       cellRenderer: (row) => <SeverityBadge severity={row.severity} />,
     },
     { label: "License", key: "license" },
-    { label: "Registry", key: "registry" },
+    {
+      label: "Registry",
+      cellRenderer: (row) => row.registry || "Unknown",
+    },
     { label: "File Path", key: "filePath" },
   ],
 };